refactor impl ToTsType to separated modules.
